test: move TestCheckVehicleCountMatch to gtfs-realtime-bindings_test.go
